Clean up tapis-redux files
