+++ src/tapis-redux/src/store/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { authenticator } from '../authenticator/reducer';
-import { systems } from '../systems/systems.reducer';
+import { systems } from '../systems/reducer';
 
 const rootReducer = combineReducers({
   authenticator,
