+++ src/tapis-redux/src/systems/selectors.ts
@@ -1,0 +1,5 @@
+import { TapisState } from '../store/rootReducer';
+
+const getSystems = (state: TapisState) => state.systems.definitions;
+
+export default getSystems;
