--- src/tapis-redux/src/systems/systems.selectors.ts
@@ -1,3 +1,0 @@
-const getSystems = (state) => state.systems.definitions;
-
-export default getSystems;
