+++ src/tapis-redux/src/types/callback.ts
@@ -1,0 +1,3 @@
+type ApiCallback<T> = (result: T | Error, ...args: any[]) => any;
+
+export default ApiCallback;
