+++ src/tapis-redux/src/types/index.ts
@@ -1,1 +1,2 @@
 export { defaultConfig, Config } from './config';
+export { default as ApiCallback } from './callback';
